chore: move action fixtures
